requirements.txt
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
segmentation-models-pytorch>=0.3.0
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
numpy>=1.24.0
opencv-python>=4.7.0
